some simulation-related functions relocated
