+++ simulation_utils.py
@@ -1,4 +1,3 @@
-import copy
 import os
 import pathlib
 import pickle
@@ -7,8 +6,7 @@ import time
 import logging
 
 from datetime import datetime
-from road_factory import get_road
-from uuid import uuid4
+# from uuid import uuid4
 
 import simulation_config as cfg
 
@@ -376,3 +374,331 @@ def scenario_finished():
     if os.path.exists(cfg.base_directory + "finished.txt"):
         return True
     return False
+
+
+def get_values(filename):
+    """
+    Extract fitness values from simulation output log files.
+    """
+    file_name = 'results/' + filename
+
+    assert os.path.exists(
+        file_name), 'The file {} does not exist!'.format(file_name)
+
+    DfC_min = 1
+    DfV_min = 1
+    DfP_min = 1
+    DfM_min = 1
+    DT_max = -1
+    traffic_lights_max = 1
+    first = True
+    distance_Max = -1
+    file_name_ex = file_name+'_ex.log'
+
+    if os.path.exists(file_name_ex):
+        with open(file_name_ex, "r") as file_handler_ex:
+            # file_handler_ex = open(file_name_ex, "r")
+            for line_ex in file_handler_ex:  # using sensors to get the data
+                if "red_light" in line_ex:
+                    logger.info("Red_light invasion")
+                    traffic_lights_max = 0
+                if "lane" in line_ex:
+                    logger.info("lane invasion")
+                    DfC_min = 0
+                if "sidewalk" in line_ex:
+                    logger.info("sidewalk invasion")
+                    DfC_min = 0
+                if "vehicle" in line_ex:
+                    logger.info("vehicle collision")
+                    DfV_min = 0
+
+    with open(file_name, "r") as file_handler:
+        for line in file_handler:
+            line_parts = line.split('>')
+            clean_line_parts = line_parts[1].replace('DfC:', '').replace(
+                'DfV:', '').replace('DfP:', '').replace('DfM:', '').replace('DT:', '')
+            double_parts = clean_line_parts.split(',')
+            DfC = float(double_parts[0])
+            DfV = float(double_parts[1])
+            DfP = float(double_parts[2])
+            DfM = float(double_parts[3])
+            DT = float(double_parts[4])
+
+            if DT < 0:
+                DT_max = 1
+                break
+
+            if first:
+                first = False
+                distance_Max = DT
+
+            DfC = 1 - (DfC / 1.15)  # normalising
+            if DfV > 1:
+                DfV = 1
+            if DfP > 1:
+                DfP = 1
+            if DfM > 1:
+                DfM = 1
+
+            distance_travelled = distance_Max - DT
+            normalised_distance_travelled = distance_travelled/distance_Max
+            if DfC < DfC_min:
+                DfC_min = DfC
+            if DfV < DfV_min:
+                DfV_min = DfV
+            if DfM < DfM_min:
+                DfM_min = DfM
+            if DfP < DfP_min:
+                DfP_min = DfP
+            if normalised_distance_travelled > DT_max:
+                DT_max = normalised_distance_travelled
+
+    return DfC_min, DfV_min, DfP_min, DfM_min, DT_max, traffic_lights_max
+
+
+# spawn points for different types of roads
+def get_road(fv, file_contents):
+    """
+    Returns a dictionary of road-related flags.
+
+    `fv` is the scenario_list while `file_contents` is an empty dictionary.
+    """
+    no_of_signals = 0
+    if fv[3] == 0:  # straight Road
+        if fv[4] == 0:  # Road ID
+            file_contents["simulator_town="] = "--simulator_town=1\n"
+            file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=29\n"
+            if (fv[5] == 0):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=31\n"
+                file_contents["goal_location="] = "--goal_location=182.202, 330.639991, 0.300000\n"
+            elif (fv[5] == 1):
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=31\n"
+                file_contents["goal_location="] = " --goal_location=234.0, 330.639991, 0.300000\n"
+            elif (fv[5] == 2):
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=31\n"
+                file_contents["goal_location="] = "--goal_location=284.0, 330.639991, 0.300000\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=35\n"
+
+        if fv[4] == 1:  # Road ID
+            file_contents["simulator_town="] = "--simulator_town=1\n"
+            if (fv[5] == 0):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=163\n"
+                file_contents["goal_location="] = "--goal_location=396.310547, 190.713867, 0.300000\n"
+            elif (fv[5] == 1):
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=64\n"
+                file_contents["goal_location="] = "--goal_location=395.959991,204.169998, 0.300000\n"
+            elif (fv[5] == 2):
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=64\n"
+                file_contents["goal_location="] = "--goal_location=395.959991,254.169998, 0.300000\n"
+            file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=65\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=130\n"
+        if fv[4] == 2:  # Road ID
+            file_contents["simulator_town="] = "--simulator_town=3\n"
+            if (fv[5] == 0):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=227\n"
+                file_contents["goal_location="] = "--goal_location=245.583176, 1.595174, 0.300000\n"
+            file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=117\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=216\n"
+            file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=119\n"
+        if fv[4] == 3:  # Road ID
+            file_contents["simulator_town="] = "--simulator_town=3\n"
+            if (fv[5] == 0):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=1\n"
+                file_contents["goal_location="] = "--goal_location=126.690155, 8.264045, 0.275307\n"
+            file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=147\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=104\n"
+            file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=148\n"
+
+    if fv[3] == 1:  # left turn Road
+        if fv[4] == 0:
+            file_contents["simulator_town="] = "--simulator_town=1\n"
+            if (fv[5] == 0):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=125\n"
+                file_contents["goal_location="] = "--goal_location=396.349457, 300.406677, 0.300000\n"
+            if (fv[5] == 1):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=125\n"
+                file_contents["goal_location="] = "--goal_location=396.449991, 230.409991, 0.300000\n"
+
+            if (fv[5] == 2):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=125\n"
+                file_contents["goal_location="] = "--goal_location=396.449991, 180.409991, 0.300000\n"
+            file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=47\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=126\n"
+
+        if fv[4] == 1:
+            file_contents["simulator_town="] = "--simulator_town=1\n"
+            if (fv[5] == 0):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=108\n"
+                file_contents["goal_location="] = "--goal_location=22.179979, 330.459991, 0.300000\n"
+            if (fv[5] == 1):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=108\n"
+                file_contents["goal_location="] = "--goal_location=22.179979, 380.459991, 0.300000\n"
+
+            if (fv[5] == 2):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=108\n"
+                file_contents["goal_location="] = "--goal_location=22.179979, 330.459991, 0.300000\n"
+            file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=123\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=54\n"
+        if fv[4] == 2:
+            no_of_signals = 1
+            file_contents["simulator_town="] = "--simulator_town=3\n"
+            if (fv[5] == 0):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=71\n"
+                file_contents["goal_location="] = "--goal_location=-84.70,-158.58,0.275307\n"
+            file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=130\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=50\n"
+            file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=133\n"
+        if fv[4] == 3:
+            no_of_signals = 1
+            file_contents["simulator_town"] = "--simulator_town=3\n"
+            if (fv[5] == 0):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=70\n"
+                file_contents["goal_location="] = "--goal_location=-88.20,-158.58,0.275307\n"
+            file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=133\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=50\n"
+            file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=130\n"
+
+    if fv[3] == 2:  # right turn Road
+
+        if fv[4] == 0:
+            file_contents["simulator_town="] = "--simulator_town=1\n"
+            if (fv[5] == 0):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=187\n"
+                file_contents["goal_location="] = "--goal_location=392.470001, 19.920038, 0.300000\n"
+            if (fv[5] == 1):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=187\n"
+                file_contents["goal_location="] = "--goal_location=392.470001, 59.920038, 0.300000\n"
+            if (fv[5] == 2):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=187\n"
+                file_contents["goal_location="] = "--goal_location=392.470001, 109.920038, 0.300000\n"
+            file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=181\n"
+            # check me
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=184\n"
+        if fv[4] == 1:
+            file_contents["simulator_town="] = "--simulator_town=1\n"
+            if (fv[5] == 0):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=57\n"
+                file_contents["goal_location="] = "--goal_location=2.009914, 295.863309, 0.300000\n"
+            if (fv[5] == 1):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=57\n"
+                file_contents["goal_location="] = "--goal_location=55.009914, 295.863309, 0.300000\n"
+            if (fv[5] == 2):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=57\n"
+                file_contents["goal_location="] = "--goal_location=108.009914, 295.863309, 0.300000\n"
+            file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=67\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=108\n"
+
+        if fv[4] == 2:
+            file_contents["simulator_town="] = "--simulator_town=5\n"
+            if (fv[5] == 0):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=244\n"
+                file_contents["goal_location="] = "--goal_location=-230.40, -84.75, 0.300000\n"
+            file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=301\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=282\n"
+            file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=293\n"
+        if fv[4] == 3:
+            file_contents["simulator_town="] = "--simulator_town=3\n"
+            if (fv[5] == 0):  # road length
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=57\n"
+                file_contents["goal_location="] = "--goal_location=-36.630997, -194.923615, 0.275307\n"
+            file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=127\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=264\n"
+            file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=128\n"
+
+    if fv[3] == 3:  # road id
+
+        if fv[4] == 0:
+            no_of_signals = 1
+            file_contents["simulator_town="] = "--simulator_town=5\n"
+            if fv[6] == 0:  # Follow Road
+                if (fv[5] == 0):  # road length
+                    file_contents["goal_location="] = "--goal_location=-220.048904, -3.915073, 0.300000\n"
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=137\n"
+                file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=59\n"
+                file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=60\n"
+            if fv[6] == 1:  # 1st exit
+                if (fv[5] == 0):  # road length
+                    file_contents["goal_location="] = "--goal_location=-184.585892, -53.541496, 0.600000\n"
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=137\n"
+                file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=59\n"
+                file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=60\n"
+            if fv[6] == 2:  # 2nd exit
+                if (fv[5] == 0):  # road length
+                    file_contents["goal_location="] = "--goal_location=-191.561127, 36.201321, 0.600000\n"
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=138\n"
+                file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=60\n"
+                file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=59\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=127\n"
+        # -----
+        if fv[4] == 1:
+            no_of_signals = 1
+            file_contents["simulator_town="] = "--simulator_town=5\n"
+            if fv[6] == 0:  # Follow Road
+                if (fv[5] == 0):  # road length
+                    file_contents["goal_location="] = "--goal_location=-188.079239, -29.370184, 0.600000\n"
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=38\n"
+                file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=34\n"
+                file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=33\n"
+
+            if fv[6] == 1:  # 1st exit
+                if (fv[5] == 0):  # road length
+                    file_contents["goal_location="] = "--goal_location=-159.701355, 6.438059, 0.300000\n"
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=37\n"
+                file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=33\n"
+                file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=34\n"
+            if fv[6] == 2:  # 2nd exit
+                if (fv[5] == 0):  # road length
+                    file_contents["goal_location="] = "--goal_location=-220.040390, -0.415084, 0.600000\n"
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=38\n"
+                file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=34\n"
+                file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=33\n"
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=131\n"
+
+        if fv[4] == 2:
+            no_of_signals = 1
+            file_contents["simulator_town="] = "--simulator_town=5\n"
+            if fv[6] == 0:  # Follow Road
+                if (fv[5] == 0):  # road length
+                    file_contents["goal_location="] = "--goal_location=-44.200703, -41.710579, 0.450000\n"
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=44\n"
+                file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=275\n"
+                file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=274\n"
+            if fv[6] == 1:  # 1st exit
+                if (fv[5] == 0):  # road length
+                    file_contents["goal_location="] = "--goal_location=0.556072, 6.047980, 0.300000\n"
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=44\n"
+                file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=275\n"
+                file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=274\n"
+            if fv[6] == 2:  # 2nd exit
+                if (fv[5] == 0):  # road length
+                    file_contents["goal_location="] = "--goal_location=-81.402634, -0.752534, 0.300000\n"
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=43\n"
+                file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=274\n"
+                file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=275\n"
+
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=206\n"
+        if fv[4] == 3:
+            no_of_signals = 1
+            file_contents["simulator_town="] = "--simulator_town=5\n"
+            if fv[6] == 0:  # Follow Road
+                if (fv[5] == 0):  # road length
+                    file_contents["goal_location="] = "--goal_location=24.741877, -52.334110, 0.300000\n"
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=75\n"
+                file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=77\n"
+                file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=78\n"
+            if fv[6] == 1:  # 1st exit
+                if (fv[5] == 0):  # road length
+                    file_contents["goal_location="] = "--goal_location=1.051966, -94.892189, 0.300000\n"
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=75\n"
+                file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=77\n"
+                file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=78\n"
+            if fv[6] == 2:  # 2nd exit
+                if (fv[5] == 0):  # road length
+                    file_contents["goal_location="] = "--goal_location=52.13, -87.77, 0.300000\n"
+                file_contents["simulator_spawn_point_index="] = "--simulator_spawn_point_index=76\n"
+                file_contents["vehicle_in_front_spawn_point="] = "--vehicle_in_front_spawn_point=78\n"
+                file_contents["vehicle_in_adjacent_spawn_point="] = "--vehicle_in_adjacent_spawn_point=77\n"
+
+            file_contents["vehicle_in_opposite_spawn_point="] = "--vehicle_in_opposite_spawn_point=218\n"
+
+    return file_contents
